nemo: implement FlatFlow-compatible GPTDataset code.
`indexed_dataset` built from `NeMoGPTDataset.__init__` is used primarily. While `doc_idx`, `sample_idx`, and `shuffle_idx`, are ignored because these are built according to the concat-then-chunk scheme.

Currently it does not support multi-epoch training by itself, for which `samples_mapping`-like virtualization needs to be implemented in the future.